requirements.txt
Django==2.1.7
python-telegram-bot==12.0.0b1
